+++ container/karaf/src/main/filtered-resources/etc/custom.properties
@@ -1183,7 +1183,6 @@ org.osgi.framework.system.packages.extra=org.apache.karaf.branding,\
         org.opennms.core.resource.db;version=${opennms.osgi.version},\
         org.opennms.core.ipc.sink.api;version=${opennms.osgi.version},\
         org.opennms.core.ipc.twin.api;version=${opennms.osgi.version},\
-        org.opennms.core.ipc.twin.common;version=${opennms.osgi.version},\
         org.opennms.core.rpc.api;version=${opennms.osgi.version},\
         org.opennms.core.rpc.echo;version=${opennms.osgi.version},\
         org.opennms.core.rpc.utils;version=${opennms.osgi.version},\
