+++ core/ipc/twin/api/src/main/java/org/opennms/core/ipc/twin/api/LocalTwinSubscriber.java
@@ -26,10 +26,9 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.core.ipc.twin.common;
+package org.opennms.core.ipc.twin.api;
 
 import com.codahale.metrics.MetricRegistry;
-import org.opennms.core.ipc.twin.api.TwinSubscriber;
 import org.opennms.core.tracing.api.TracerRegistry;
 
 public interface LocalTwinSubscriber extends TwinSubscriber {
