+++ core/ipc/twin/api/src/main/java/org/opennms/core/ipc/twin/api/TwinTracker.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.core.ipc.twin.common;
+package org.opennms.core.ipc.twin.api;
 
 import java.util.Objects;
 import java.util.UUID;
