+++ core/ipc/twin/grpc/publisher/src/main/java/org/opennms/core/ipc/twin/grpc/publisher/GrpcTwinPublisher.java
@@ -40,9 +40,9 @@ import org.opennms.core.grpc.common.GrpcIpcServer;
 import org.opennms.core.grpc.common.GrpcIpcUtils;
 import org.opennms.core.ipc.twin.api.TwinStrategy;
 import org.opennms.core.ipc.twin.common.AbstractTwinPublisher;
-import org.opennms.core.ipc.twin.common.LocalTwinSubscriber;
-import org.opennms.core.ipc.twin.common.TwinRequest;
-import org.opennms.core.ipc.twin.common.TwinUpdate;
+import org.opennms.core.ipc.twin.api.LocalTwinSubscriber;
+import org.opennms.core.ipc.twin.api.TwinRequest;
+import org.opennms.core.ipc.twin.api.TwinUpdate;
 import org.opennms.core.ipc.twin.grpc.common.MinionHeader;
 import org.opennms.core.ipc.twin.grpc.common.OpenNMSTwinIpcGrpc;
 import org.opennms.core.ipc.twin.model.TwinRequestProto;
