+++ core/ipc/twin/grpc/subscriber/src/main/java/org/opennms/core/ipc/twin/grpc/subscriber/GrpcTwinSubscriber.java
@@ -36,8 +36,8 @@ import io.grpc.ManagedChannel;
 import io.grpc.stub.StreamObserver;
 import org.opennms.core.grpc.common.GrpcIpcUtils;
 import org.opennms.core.ipc.twin.common.AbstractTwinSubscriber;
-import org.opennms.core.ipc.twin.common.TwinRequest;
-import org.opennms.core.ipc.twin.common.TwinUpdate;
+import org.opennms.core.ipc.twin.api.TwinRequest;
+import org.opennms.core.ipc.twin.api.TwinUpdate;
 import org.opennms.core.ipc.twin.grpc.common.MinionHeader;
 import org.opennms.core.ipc.twin.grpc.common.OpenNMSTwinIpcGrpc;
 import org.opennms.core.ipc.twin.model.TwinRequestProto;
