+++ core/ipc/twin/jms/publisher/src/main/java/org/opennms/core/ipc/twin/jms/publisher/JmsTwinPublisher.java
@@ -47,9 +47,9 @@ import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.jms.JmsEndpoint;
 import org.opennms.core.ipc.twin.api.TwinStrategy;
 import org.opennms.core.ipc.twin.common.AbstractTwinPublisher;
-import org.opennms.core.ipc.twin.common.LocalTwinSubscriber;
-import org.opennms.core.ipc.twin.common.TwinRequest;
-import org.opennms.core.ipc.twin.common.TwinUpdate;
+import org.opennms.core.ipc.twin.api.LocalTwinSubscriber;
+import org.opennms.core.ipc.twin.api.TwinRequest;
+import org.opennms.core.ipc.twin.api.TwinUpdate;
 import org.opennms.core.ipc.twin.model.TwinResponseProto;
 import org.opennms.core.tracing.api.TracerRegistry;
 import org.opennms.core.tracing.util.TracingInfoCarrier;
