+++ core/ipc/twin/jms/subscriber/src/main/java/org/opennms/core/ipc/twin/jms/subscriber/JmsTwinSubscriber.java
@@ -42,8 +42,8 @@ import org.apache.camel.component.jms.JmsEndpoint;
 import org.apache.camel.impl.DefaultCamelContext;
 import org.apache.camel.spi.Synchronization;
 import org.opennms.core.ipc.twin.common.AbstractTwinSubscriber;
-import org.opennms.core.ipc.twin.common.TwinRequest;
-import org.opennms.core.ipc.twin.common.TwinUpdate;
+import org.opennms.core.ipc.twin.api.TwinRequest;
+import org.opennms.core.ipc.twin.api.TwinUpdate;
 import org.opennms.core.ipc.twin.model.TwinRequestProto;
 import org.opennms.core.tracing.api.TracerRegistry;
 import org.opennms.core.utils.SystemInfoUtils;
