+++ core/ipc/twin/kafka/subscriber/src/main/java/org/opennms/core/ipc/twin/kafka/subscriber/KafkaTwinSubscriber.java
@@ -44,8 +44,8 @@ import org.apache.kafka.common.serialization.StringSerializer;
 import org.opennms.core.ipc.common.kafka.KafkaConfigProvider;
 import org.opennms.core.ipc.common.kafka.Utils;
 import org.opennms.core.ipc.twin.common.AbstractTwinSubscriber;
-import org.opennms.core.ipc.twin.common.TwinRequest;
-import org.opennms.core.ipc.twin.common.TwinUpdate;
+import org.opennms.core.ipc.twin.api.TwinRequest;
+import org.opennms.core.ipc.twin.api.TwinUpdate;
 import org.opennms.core.ipc.twin.kafka.common.KafkaConsumerRunner;
 import org.opennms.core.ipc.twin.kafka.common.Topic;
 import org.opennms.core.ipc.twin.model.TwinRequestProto;
